Rename directory "ner" to "ner_lib", because there was name conflict between "ner.py" and this directory (for Python)
