+++ ner_lib/ner_vars.py
